feat: moved structure to match 9.2
