--- instrumentation/jetty-httpclient/jetty-httpclient-9.0/library/jetty-httpclient-9.0-library.gradle
@@ -1,25 +1,0 @@
-apply from: "$rootDir/gradle/instrumentation-library.gradle"
-apply plugin: "net.ltgt.errorprone"
-
-def jettyVers = [:]
-
-//Jetty client 9.1 is the best starting point for the v9.x series, v9.0.x is incompatible
-jettyVers.base9 = '9.1.0.v20131115'
-//jettyVers.base9 = '9.0.0.v20130308'
-jettyVers.useInTests9 = '9.3.28.v20191105'
-
-dependencies {
-  library "org.eclipse.jetty:jetty-client:${jettyVers.useInTests9}"
-  testLibrary "org.eclipse.jetty:jetty-client:${jettyVers.useInTests9}"
-  testImplementation project(':instrumentation:jetty-httpclient::jetty-httpclient-9.0:testing')
-
-  implementation deps.slf4j
-
-  testImplementation("org.eclipse.jetty:jetty-server:${jettyVers.useInTests9}") {
-    exclude group: 'org.eclipse.jetty', module: 'jetty-client'
-  }
-//  testLibrary "org.eclipse.jetty:jetty-client:${jettyVers.useInTests9}"
-
-
-}
-
