--- instrumentation/jetty-httpclient/jetty-httpclient-9.0/testing/jetty-httpclient-9.0-testing.gradle
@@ -1,20 +1,0 @@
-apply from: "$rootDir/gradle/java.gradle"
-def jettyVers = [:]
-jettyVers.base9 = '9.1.0.v20131115'
-jettyVers.useInTests9 = '9.3.28.v20191105'
-dependencies {
-  api(project(':testing-common')) {
-    exclude group: 'org.eclipse.jetty', module: 'jetty-client'
-    exclude group: 'org.eclipse.jetty', module: 'jetty-server'
-  }
-
-
-  api "org.eclipse.jetty:jetty-client:${jettyVers.useInTests9}"
-
-  implementation deps.guava
-  implementation deps.junitApi
-
-  implementation deps.groovy
-  implementation deps.opentelemetryApi
-  implementation deps.spock
-}
