+++ instrumentation/jetty-httpclient/jetty-httpclient-9.2/javaagent/jetty-httpclient-9.0-javaagent.gradle
@@ -1,0 +1,27 @@
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+muzzle {
+  pass {
+    group = "org.eclipse.jetty"
+    module = 'jetty-client'
+    versions = "[9.1,10)"
+    assertInverse = true
+  }
+}
+
+def jettyVers = [:]
+
+//Jetty client 9.2 is the best starting point, HttpClient.send() is stable there
+jettyVers.base9 = '9.2.0.v20140526'
+jettyVers.useInTests9 = '9.2.0.v20140526'
+
+dependencies {
+  implementation project(':instrumentation:jetty-httpclient:jetty-httpclient-9.0:library')
+
+  library "org.eclipse.jetty:jetty-client:${jettyVers.base9}"
+
+  testImplementation project(':instrumentation:jetty-httpclient:jetty-httpclient-9.0:testing')
+  testImplementation("org.eclipse.jetty:jetty-server:${jettyVers.useInTests9}") {
+    exclude group: 'org.eclipse.jetty', module: 'jetty-client'
+  }
+}
