+++ instrumentation/jetty-httpclient/jetty-httpclient-9.2/javaagent/src/test/groovy/io/opentelemetry/javaagent/instrumentation/jetty/httpclient/v9_2/JettyHttpClient9AgentTest.groovy
@@ -1,0 +1,24 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.jetty.httpclient.v9_2
+
+
+import io.opentelemetry.instrumentation.test.AgentTestTrait
+import org.eclipse.jetty.client.HttpClient
+import org.eclipse.jetty.util.ssl.SslContextFactory
+
+class JettyHttpClient9AgentTest extends AbstractJettyClient9Test implements AgentTestTrait {
+
+  @Override
+  HttpClient createStandardClient() {
+    return new HttpClient()
+  }
+
+  @Override
+  HttpClient createHttpsClient(SslContextFactory sslContextFactory) {
+    return new HttpClient(sslContextFactory)
+  }
+}
