+++ settings.gradle
@@ -185,9 +185,9 @@ include ':instrumentation:jedis:jedis-3.0:javaagent'
 include ':instrumentation:jetty:jetty-8.0:javaagent'
 include ':instrumentation:jetty:jetty-11.0:javaagent'
 include ':instrumentation:jetty:jetty-common:javaagent'
-include ':instrumentation:jetty-httpclient:jetty-httpclient-9.0:javaagent'
-include ':instrumentation:jetty-httpclient:jetty-httpclient-9.0:library'
-include ':instrumentation:jetty-httpclient:jetty-httpclient-9.0:testing'
+include ':instrumentation:jetty-httpclient:jetty-httpclient-9.2:javaagent'
+include ':instrumentation:jetty-httpclient:jetty-httpclient-9.2:library'
+include ':instrumentation:jetty-httpclient:jetty-httpclient-9.2:testing'
 include ':instrumentation:jms-1.1:javaagent'
 include ':instrumentation:jms-1.1:javaagent-unit-tests'
 include ':instrumentation:jsf:jsf-common:library'
